+++ src/main/java/es/iesjandula/hotelv1/gestionhotel/Enum/MetodoPago.java
@@ -1,21 +1,22 @@
 //Paquete.
-package es.iesjandula.hotelv1.gestionhotel.model;
+package es.iesjandula.hotelv1.gestionhotel.Enum;
 
 /**
- * Definición de la clase MetodoPago.
+ * Definición de la clase enumerada MetodoPago.
  */
 public enum MetodoPago {
+    //Definimos los posibles estados del Método de pago.
     TARJETA_CREDITO("Tarjeta de Credito"),
     TARJETA_DEBITO("Tarjeta de Debito"),
     EFECTIVO("Efectivo"),
     TRANSFERENCIA("Transferencia"),
     PAYPAL("Paypal"),;
 
-    //Atributo
+    //Atributo nombre del estado.
     private final String nombre;
 
     /**
-     * Constructor de la clase
+     * Constructor de la clase enumerada MetodoPago con el siguiente parámetro.
      * @param nombre
      */
     MetodoPago(String nombre) {
@@ -28,9 +29,9 @@ public enum MetodoPago {
     }
 
 
-    //Método toString sobreescrito.
+    //Método toString sobrescrito.
     @Override
     public String toString() {
         return getNombre();
-    }
+    } // Devuelve el nombre del estado, utilizando el getter
 }
