--- src/main/java/es/iesjandula/hotelv1/gestionhotel/model/EstadoReserva.java
@@ -1,8 +1,0 @@
-package es.iesjandula.hotelv1.gestionhotel.model;
-
-public enum EstadoReserva {
-    PENDIENTE,
-    CONFIRMADA,
-    CANCELADA,
-    FINALIZADA
-}
