+++ src/main/java/es/iesjandula/hotelv1/gestionhotel/model/Pago.java
@@ -2,9 +2,10 @@
 package es.iesjandula.hotelv1.gestionhotel.model;
 
 //Importaciones.
+import es.iesjandula.hotelv1.gestionhotel.Enum.EstadoPago;
+import es.iesjandula.hotelv1.gestionhotel.Enum.MetodoPago;
 import jakarta.persistence.*;
 import java.util.Date;
-import java.util.List;
 
 /**
  * Definición de la entidad pago.
@@ -12,12 +13,13 @@ import java.util.List;
 @Entity
 @Table(name="Pago")
 public class Pago {
+
+    //Atributos.
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     private long id;
     private double monto;
     private Date fecha;
-
     @Enumerated(EnumType.STRING)
     private MetodoPago metodo;
     @Enumerated(EnumType.STRING)
@@ -36,6 +38,15 @@ public class Pago {
     public Pago() {
     }
 
+    /**
+     * Constructor de la clase (Entidad) Pago con los siguientes parámetros.
+     * @param id
+     * @param monto
+     * @param fecha
+     * @param metodo
+     * @param estado
+     * @param reserva
+     */
 
     public Pago(long id, double monto, Date fecha, MetodoPago metodo, EstadoPago estado, Reserva reserva) {
         this.id = id;
@@ -46,6 +57,7 @@ public class Pago {
         this.reserva = reserva;
     }
 
+    //Métodos de acceso.
     public long getId() {
         return id;
     }
@@ -94,6 +106,7 @@ public class Pago {
         this.reserva = reserva;
     }
 
+    //Método toString() sobrescrito.
     @Override
     public String toString() {
         return String.format(
@@ -106,7 +119,5 @@ public class Pago {
                 (reserva != null ? reserva.toString() : "null")
         );
     }
-
-
 
 }
